Moving Utils.js to a higher scope
