+++ lib/lambda/UtilsTest.ts
@@ -1,4 +1,5 @@
-import { AbstractRoleApi, IncomingPayload, LambdaProxyIntegrationResponse, OutgoingBody } from "../../role/AbstractRole";
+import { AbstractRoleApi, IncomingPayload, LambdaProxyIntegrationResponse, OutgoingBody } from "../role/AbstractRole";
+import { deepEqual } from "./Utils";
 
 /**
  * Helper class used in unit tests to keep track of and reset calls to mocks.
