+++ lib/lambda/functions/sys-admin/view/HtmlTableView.ts
@@ -1,4 +1,4 @@
-import { viewHtml } from "../../Utils";
+import { viewHtml } from "../../../Utils";
 import { AbstractTableView } from "./AbstractTableView";
 
 /**
